requirements.txt
attrs==23.2.0
boxsdk==3.9.2
certifi==2024.2.2
charset-normalizer==3.3.2
exceptiongroup==1.2.0
h11==0.14.0
idna==3.6
outcome==1.3.0.post0
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.1
requests==2.31.0
requests-toolbelt==1.0.0
selenium==4.17.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
trio==0.24.0
trio-websocket==0.11.1
typing_extensions==4.9.0
urllib3==2.2.0
wsproto==1.2.0
